server/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg[binary]>=3.1,<4.0
drf-spectacular>=0.29.0
